Removing casts and imports now not needed after generics applied.
git-svn-id: https://svn.apache.org/repos/asf/cayenne/main/trunk@600160 13f79535-47bb-0310-9956-ffa450edef68

+++ framework/cayenne-jdk1.5-unpublished/src/main/java/org/apache/cayenne/ObjectContextChangeLog.java
@@ -91,7 +91,7 @@ class ObjectContextChangeLog {
     }
 
     GraphDiff getDiffsAfterMarker(String markerTag) {
-        Integer pos = (Integer) markers.get(markerTag);
+        Integer pos = markers.get(markerTag);
         int marker = (pos == null) ? -1 : pos.intValue();
         if (marker < 0) {
             throw new IllegalStateException("No marked position for tag '"
@@ -122,7 +122,7 @@ class ObjectContextChangeLog {
     }
 
     int sizeAfterMarker(String markerTag) {
-        Integer pos = (Integer) markers.get(markerTag);
+        Integer pos = markers.get(markerTag);
         int marker = (pos == null) ? -1 : pos.intValue();
         if (marker < 0) {
             throw new IllegalStateException("No marked position for tag '"
