+++ framework/cayenne-jdk1.5-unpublished/src/main/java/org/apache/cayenne/wocompat/parser/ParserTokenManager.java
@@ -20,9 +20,6 @@
 
 
 package org.apache.cayenne.wocompat.parser;
-import java.io.*;
-import java.util.*;
-import org.apache.cayenne.wocompat.PlistDataStructureFactory;
 
 public class ParserTokenManager implements ParserConstants
 {
@@ -864,9 +861,9 @@ void MoreLexicalActions()
          jjimageLen = 0;
         int oct;
         len = image.length();
-        oct = ((int)(image.charAt(len-1)-'0')) +
-              8 * ((int)(image.charAt(len-2)-'0')) +
-              64 * ((int)(image.charAt(len-3)-'0'));
+        oct = ((image.charAt(len-1)-'0')) +
+              8 * ((image.charAt(len-2)-'0')) +
+              64 * ((image.charAt(len-3)-'0'));
         image.setCharAt(len-4, (char)oct);
         image.setLength(len-3);
          break;
@@ -879,7 +876,7 @@ void MoreLexicalActions()
         int dig,c, base=1,hex=0;
         len = image.length();
         for(int i=0; i<4; ++i) {
-          c=(int)image.charAt(len-(i+1));
+          c=image.charAt(len-(i+1));
           if (c<='9') c-='0';
             else if (c<='F') c = (c -'A') + 10;
               else c =(c -'a') + 10;
